Create abstract class for compliance policy enforced settings
