+++ plugins/Resolution/Settings/ScreenResolutionDetectionDisabled.php
@@ -3,51 +3,19 @@
 namespace Piwik\Plugins\Resolution\Settings;
 
 use Piwik\Piwik;
+use Piwik\Plugins\PrivacyManager\Settings\CompliancePolicyEnforcedSetting;
 use Piwik\Policy\CnilPolicy;
-use Piwik\Settings\Interfaces\PolicyComparisonInterface;
-use Piwik\Settings\Interfaces\SettingValueInterface;
-use Piwik\Settings\Interfaces\Traits\PolicyComparisonTrait;
 
-/**
- * @implements PolicyComparisonInterface<bool>
- * @implements SettingValueInterface<bool>
- */
-class ScreenResolutionDetectionDisabled implements
-    PolicyComparisonInterface,
-    SettingValueInterface
+class ScreenResolutionDetectionDisabled extends CompliancePolicyEnforcedSetting
 {
-    /** @use PolicyComparisonTrait<bool> */
-    use PolicyComparisonTrait;
-
-    /** @var bool */
-    private $value;
-
-    protected function __construct(bool $value)
-    {
-        $this->value = $value;
-    }
-
-    protected static function compareStrictness($value1, $value2)
-    {
-        return $value1 || $value2;
-    }
-
     public static function getTitle(): string
     {
         return Piwik::translate('Resolution_ScreenResolutionDetectionDisabled');
     }
 
-    public static function getInstance(?int $idSite = null): self
+    public static function getComplianceRequirementNote(?int $idSite = null): string
     {
-        $values = self::getPolicyRequiredValues($idSite);
-        $strictest = (bool) self::getStrictestValueFromArray($values);
-
-        return new self($strictest);
-    }
-
-    public static function getInlineHelp(): string
-    {
-        return '';
+        return Piwik::translate('Resolution_ScreenResolutionDetectionDisabledRequirementNote');
     }
 
     public static function getPolicyRequirements(): array
@@ -55,27 +23,5 @@ class ScreenResolutionDetectionDisabled implements
         return [
             CnilPolicy::class => true,
         ];
-    }
-
-    public static function isCompliant(string $policy, ?int $idSite = null): bool
-    {
-        $policyValues = self::getPolicyRequirements();
-        if (!array_key_exists($policy, $policyValues)) {
-            return true;
-        }
-
-        $currentValue = self::getInstance($idSite)->getValue();
-
-        return $currentValue === $policyValues[$policy];
-    }
-
-    public function getValue()
-    {
-        return $this->value;
-    }
-
-    public static function getComplianceRequirementNote(?int $idSite = null): string
-    {
-        return Piwik::translate('Resolution_ScreenResolutionDetectionDisabledRequirementNote');
     }
 }
